Convert type definitions back to ES module definitions, add more entry points
