+++ package.json
@@ -5,13 +5,12 @@
   "main": "web.cjs.js",
   "module": "web.js",
   "react-native": "native.js",
-  "types": "types/index.d.ts",
   "private": true,
   "sideEffects": false,
   "scripts": {
     "prebuild": "rimraf dist",
     "build": "npm-run-all --parallel copy rollup",
-    "copy": "copyfiles package.json readme.md LICENSE.md types dist && json -I -f dist/package.json -e \"this.private=false; this.devDependencies=undefined; this.scripts=undefined; this.husky=undefined; this.prettier=undefined; this.jest=undefined; this['lint-staged'] = undefined;\"",
+    "copy": "copyfiles -f package.json readme.md LICENSE.md \"types/*\" dist && json -I -f dist/package.json -e \"this.private=false; this.devDependencies=undefined; this.scripts=undefined; this.husky=undefined; this.prettier=undefined; this.jest=undefined; this['lint-staged'] = undefined;\"",
     "rollup": "rollup -c",
     "prepare": "npm run build",
     "test": "jest",
@@ -89,6 +88,7 @@
     "immer-wieder": "^1.1.5",
     "jest": "^23.6.0",
     "json": "^9.0.6",
+    "konva": "^2.6.0",
     "lint-staged": "^7.3.0",
     "lorem-ipsum": "^1.0.6",
     "mock-raf": "^1.0.0",
@@ -99,6 +99,7 @@
     "react-dom": "16.7.0-alpha.2",
     "react-feather": "^1.1.4",
     "react-helmet": "^5.2.0",
+    "react-konva": "^16.6.31",
     "react-test-renderer": "^16.5.2",
     "react-testing-library": "^5.2.0",
     "react-transition-group": "^2.5.0",
