+++ types/addons.d.ts
@@ -1,32 +1,28 @@
-/// <reference types="./react-spring"/>
+import { Ref, PureComponent } from 'react'
+import { SpringConfig } from './universal'
 
-declare module 'react-spring/addons' {
-  import { Ref, PureComponent } from 'react'
-  import { SpringConfig } from 'react-spring'
+interface ParallaxProps {
+  pages: number
 
-  interface ParallaxProps {
-    pages: number
+  config?: SpringConfig | ((key: string) => SpringConfig)
 
-    config?: SpringConfig | ((key: string) => SpringConfig)
+  scrolling?: boolean
 
-    scrolling?: boolean
+  horizontal?: boolean
 
-    horizontal?: boolean
-
-    ref?: Ref<Parallax>
-  }
-
-  export class Parallax extends PureComponent<ParallaxProps> {
-    scrollTo: (offset: number) => void
-  }
-
-  interface ParallaxLayerProps {
-    factor?: number
-
-    offset?: number
-
-    speed?: number
-  }
-
-  export class ParallaxLayer extends PureComponent<ParallaxLayerProps> {}
+  ref?: Ref<Parallax>
 }
+
+export class Parallax extends PureComponent<ParallaxProps> {
+  scrollTo: (offset: number) => void
+}
+
+interface ParallaxLayerProps {
+  factor?: number
+
+  offset?: number
+
+  speed?: number
+}
+
+export class ParallaxLayer extends PureComponent<ParallaxLayerProps> {}
