+++ types/native.d.ts
@@ -1,0 +1,1 @@
+export * from './universal'
